move Plane & Sphere in Scene module
