+++ Engine/Physics/test/test_Collisions.cpp
@@ -3,6 +3,7 @@
 #include <gtest/gtest.h>
 
 using namespace Stone::Physics;
+using namespace Stone::Scene;
 
 TEST(PhysicsEngineTest, UpdateSpherePosition) {
 	Sphere sphere({0.0f, 0.0f, 0.0f}, 1.0f);
